:+1: Use `Worker` to isolate each connection
