+++ deno.jsonc
@@ -4,7 +4,7 @@
     "test": "deno test -A --parallel --shuffle --doc",
     "test:coverage": "deno task test --coverage=.coverage",
     "check": "deno check ./**/*.ts",
-    "coverage": "deno coverage .coverage --exclude=cli.ts --exclude=testdata/",
+    "coverage": "deno coverage .coverage --exclude=cli.ts --exclude=worker.ts --exclude=testdata/",
     "upgrade": "deno run -q -A https://deno.land/x/molt@0.14.2/cli.ts ./**/*.ts",
     "upgrade:commit": "deno task -q upgrade --commit --prefix :package: --pre-commit=fmt"
   }
